requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose>=3.2.0
passlib[bcrypt]>=1.7.4

